升级倒 esp-idf 4.3

+++ filesystem/root/etc/init.d/app.js
@@ -11,6 +11,9 @@ exports.autoBoot = function() {
     }
 }
 exports.setAutoBootPath = function(path) {
+    if(!path||!beapi.fs.existsSync(path)) {
+        throw new Error("path not exists")
+    }
     let json = JSON.load("/etc/config/boot.json")
     json["active-app"] = path
     beapi.fs.writeFileSync("/etc/config/boot.json", JSON.stringify(json,null,4))
